+++ experiments/cvpr-2024/vcd_figs_for_abst.py
@@ -3,6 +3,7 @@ import time
 import jax.numpy as jnp
 import matplotlib.pyplot as plt
 import mbirjax
+import mbirjax.plot_utils as pu
 
 
 def display_slices_for_abstract( recon1, recon2, recon3, labels) :
@@ -99,7 +100,7 @@ if __name__ == "__main__":
     granularity_sequences = [granularity_sequence_gd, granularity_sequence_vcd, granularity_sequence_cd]
     losses = [fm_rmse_gd, fm_rmse_vcd, fm_rmse_cd]
     labels = ['Gradient Descent', 'Vectorized Coordinate Descent', 'Coordinate Descent']
-    mbirjax.plot_granularity_and_loss(granularity_sequences, losses, labels, granularity_ylim=(0, 256), loss_ylim=(0.1, 15))
+    pu.plot_granularity_and_loss(granularity_sequences, losses, labels, granularity_ylim=(0, 256), loss_ylim=(0.1, 15))
 
 
     # Generate sequence of partition images for Figure 1
@@ -107,4 +108,4 @@ if __name__ == "__main__":
     partitions_fig = parallel_model.gen_set_of_voxel_partitions()
 
     # Plot the set of partitions
-    mbirjax.debug_plot_partitions(partitions=partitions_fig, num_recon_rows=32, num_recon_cols=32)
+    pu.debug_plot_partitions(partitions=partitions_fig, num_recon_rows=32, num_recon_cols=32)
